Added Support for Unsigned 64-bit Integers
Summary: Updated atItem compare method to accomodate 64-bit addresses.

Test Plan: Compiled on Linux

Reviewers: szielinski, rlouden, martin

Reviewed By: szielinski, rlouden, martin

Differential Revision: http://cr.irl.ucf.edu/D398

+++ foundation/atItem.c++
@@ -27,6 +27,6 @@ bool atItem::equals(atItem * otherItem)
 int atItem::compare(atItem * otherItem)
 {
    // Return the difference between the items' pointer values
-   return (int) otherItem - (int) this;
+   return (long) otherItem - (long) this;
 }
